backend/IA/requirements.txt
fastapi
uvicorn
langchain
langchain-openai  
openai           
python-dotenv       
playwright